Give Features and Assets their own pages; turn project overview into a dashboard
- Add features.index page (project/{project}/features) with the full
  feature list, create, and drag-reorder UI moved out of project.show
- Add assets.index page (project/{project}/assets) wrapping the existing
  project-assets-panel component
- Rewrite project.show as a true overview: project header + 6-card nav
  grid linking to Features, Stories, Plans, Approvals, Runs, Assets
- Fix project description to render via <x-markdown> instead of {{ }}
- Add Assets and Features sidebar links; Features now points to the
  dedicated features.index route instead of the buried #features anchor
- Update ProjectShowTest to call features-related actions on the
  features.index component
