+++ resources/views/pages/projects/⚡show.blade.php
@@ -1,9 +1,6 @@
 <?php
 
-use App\Enums\FeatureStatus;
-use App\Models\Feature;
 use App\Models\Project;
-use App\Services\Ordering\PositionReorderer;
 use Illuminate\Support\Facades\Auth;
 use Livewire\Attributes\Computed;
 use Livewire\Attributes\Title;
@@ -12,12 +9,6 @@ use Livewire\Component;
 
 new #[Title('Project')] class extends Component {
     public int $project_id;
-
-    #[Validate('required|string|max:255')]
-    public string $newFeatureName = '';
-
-    #[Validate('nullable|string')]
-    public string $newFeatureDescription = '';
 
     public bool $editing = false;
 
@@ -49,32 +40,6 @@ new #[Title('Project')] class extends Component {
             ->whereIn('id', Auth::user()->accessibleProjectIds())
             ->with('team.workspace')
             ->find($this->project_id);
-    }
-
-    #[Computed]
-    public function features()
-    {
-        return $this->project
-            ? $this->project->features()
-                ->withCount('stories')
-                ->orderBy('position')
-                ->orderBy('id')
-                ->get()
-            : collect();
-    }
-
-    /**
-     * @param  array<int, int>  $orderedIds  Feature IDs in the new visual order.
-     */
-    public function reorderFeatures(array $orderedIds): void
-    {
-        $project = $this->project;
-        abort_unless($project, 404);
-        abort_unless(Auth::user()->canApproveInProject($project), 403);
-
-        app(PositionReorderer::class)->reorder('features', 'project_id', (int) $project->id, $orderedIds);
-
-        unset($this->features);
     }
 
     public function canEditProject(): bool
@@ -128,25 +93,6 @@ new #[Title('Project')] class extends Component {
         unset($this->project);
     }
 
-    public function createFeature(): void
-    {
-        $project = $this->project;
-        abort_unless($project, 404);
-        abort_unless(Auth::user()->canApproveInProject($project), 403);
-
-        $this->validate(['newFeatureName' => 'required|string|max:255']);
-
-        Feature::create([
-            'project_id' => $project->id,
-            'name' => $this->newFeatureName,
-            'description' => $this->newFeatureDescription ?: null,
-            'status' => FeatureStatus::Proposed,
-        ]);
-
-        $this->reset(['newFeatureName', 'newFeatureDescription']);
-        unset($this->features);
-    }
-
     public function deleteProject(): void
     {
         $project = $this->project;
@@ -180,8 +126,6 @@ new #[Title('Project')] class extends Component {
     @if (! $this->project)
         <flux:text class="text-zinc-500">{{ __('Project not found.') }}</flux:text>
     @else
-        @php $canManageFeatures = Auth::user()->canApproveInProject($this->project); @endphp
-
         <section data-section="project-header" class="flex flex-col gap-2">
             @if ($editing)
                 <flux:input wire:model="editName" :label="__('Name')" required />
@@ -205,12 +149,16 @@ new #[Title('Project')] class extends Component {
                     @endif
                 </div>
                 @if ($this->project->description)
-                    <flux:text class="text-zinc-600 dark:text-zinc-400">{{ $this->project->description }}</flux:text>
+                    <x-markdown :content="$this->project->description" class="text-zinc-600 dark:text-zinc-400" />
                 @endif
             @endif
         </section>
 
-        <section class="grid gap-3 md:grid-cols-3" data-section="project-workspaces">
+        <section class="grid gap-3 md:grid-cols-3" data-section="project-nav">
+            <a href="{{ route('features.index', ['project' => $this->project->id]) }}" wire:navigate class="rounded-xl border border-zinc-200 p-4 hover:bg-zinc-50 dark:border-zinc-700 dark:hover:bg-zinc-900">
+                <div class="text-xs uppercase tracking-wide text-zinc-500">{{ __('Features') }}</div>
+                <div class="mt-2 text-sm text-zinc-600 dark:text-zinc-400">{{ __('Organise work into product-level capabilities.') }}</div>
+            </a>
             <a href="{{ route('stories.index', ['project' => $this->project->id]) }}" wire:navigate class="rounded-xl border border-zinc-200 p-4 hover:bg-zinc-50 dark:border-zinc-700 dark:hover:bg-zinc-900">
                 <div class="text-xs uppercase tracking-wide text-zinc-500">{{ __('Stories') }}</div>
                 <div class="mt-2 text-sm text-zinc-600 dark:text-zinc-400">{{ __('Browse product contracts and their current plans.') }}</div>
@@ -223,103 +171,35 @@ new #[Title('Project')] class extends Component {
                 <div class="text-xs uppercase tracking-wide text-zinc-500">{{ __('Approvals') }}</div>
                 <div class="mt-2 text-sm text-zinc-600 dark:text-zinc-400">{{ __('Review story contracts and current plans in separate queues.') }}</div>
             </a>
+            <a href="{{ route('runs.index', ['project' => $this->project->id]) }}" wire:navigate class="rounded-xl border border-zinc-200 p-4 hover:bg-zinc-50 dark:border-zinc-700 dark:hover:bg-zinc-900">
+                <div class="text-xs uppercase tracking-wide text-zinc-500">{{ __('Runs') }}</div>
+                <div class="mt-2 text-sm text-zinc-600 dark:text-zinc-400">{{ __('Track AI agent execution history.') }}</div>
+            </a>
+            <a href="{{ route('assets.index', ['project' => $this->project->id]) }}" wire:navigate class="rounded-xl border border-zinc-200 p-4 hover:bg-zinc-50 dark:border-zinc-700 dark:hover:bg-zinc-900">
+                <div class="text-xs uppercase tracking-wide text-zinc-500">{{ __('Assets') }}</div>
+                <div class="mt-2 text-sm text-zinc-600 dark:text-zinc-400">{{ __('Reference material shared across all stories in this project.') }}</div>
+            </a>
         </section>
 
-        <section id="features" class="flex flex-col gap-3">
-            <div class="flex items-center justify-between gap-2">
-                <flux:heading size="lg">{{ __('Features') }}</flux:heading>
-                @if ($canManageFeatures)
-                    <flux:modal.trigger name="new-feature-modal">
-                        <flux:button variant="primary">{{ __('+ New feature') }}</flux:button>
-                    </flux:modal.trigger>
-                @endif
-            </div>
-
-            @php $canReorderFeatures = $canManageFeatures && $this->features->count() > 1; @endphp
-            <div
-                class="flex flex-col gap-3"
-                @if ($canReorderFeatures)
-                    x-data
-                    x-sortable="$wire.reorderFeatures"
-                @endif
-                data-features-list
-            >
-                @forelse ($this->features as $feature)
-                    <div
-                        wire:key="feature-{{ $feature->id }}"
-                        @if ($canReorderFeatures) data-sortable-id="{{ $feature->id }}" @endif
-                    >
-                        <flux:card>
-                            <div class="flex items-center gap-3">
-                                @if ($canReorderFeatures)
-                                    <button
-                                        type="button"
-                                        data-sortable-handle
-                                        class="flex flex-none cursor-grab items-center text-zinc-400 hover:text-zinc-600 active:cursor-grabbing dark:hover:text-zinc-300"
-                                        aria-label="{{ __('Drag to reorder') }}"
-                                        title="{{ __('Drag to reorder') }}"
-                                    >
-                                        <flux:icon.bars-3 class="size-5" />
-                                    </button>
-                                @endif
-                                <div class="min-w-0 flex-1">
-                                    <a href="{{ route('features.show', ['project' => $this->project->id, 'feature' => $feature->id]) }}" wire:navigate>
-                                        <flux:heading>{{ $feature->name }}</flux:heading>
-                                    </a>
-                                    @if ($feature->description)
-                                        <x-markdown :content="$feature->description" class="mt-1" />
-                                    @endif
-                                </div>
-                                <flux:badge>{{ $feature->stories_count }} {{ __('stories') }}</flux:badge>
-                            </div>
-                        </flux:card>
-                    </div>
-                @empty
-                    <flux:text class="text-zinc-500">{{ __('No features yet.') }}</flux:text>
-                @endforelse
-            </div>
-        </section>
-
-        <livewire:pages::context-items.project-assets-panel
-            :project-id="$this->project->id"
-            :key="'project-assets-panel-'.$this->project->id"
-        />
-
-        @if ($canManageFeatures)
-            <flux:modal name="new-feature-modal" class="md:w-96">
-                <form wire:submit.prevent="createFeature" class="flex flex-col gap-4">
-                    <flux:heading size="lg">{{ __('New feature') }}</flux:heading>
-                    <flux:input wire:model="newFeatureName" :label="__('Name')" required />
-                    <flux:textarea wire:model="newFeatureDescription" :label="__('Description (optional)')" rows="2" />
+        @if ($this->canDeleteProject())
+            <flux:modal name="delete-project-modal" class="md:w-96">
+                <div class="flex flex-col gap-4">
+                    <flux:heading size="lg">{{ __('Delete project?') }}</flux:heading>
+                    <flux:text>{{ __('This permanently removes the project, its features, stories, tasks, subtasks, approvals, and repo attachments. This cannot be undone.') }}</flux:text>
+                    <flux:input
+                        wire:model="deleteConfirmationName"
+                        :label="__('Type the project name to confirm')"
+                        :placeholder="$this->project->name"
+                        required
+                    />
                     <div class="flex justify-end gap-2">
                         <flux:modal.close>
                             <flux:button type="button" variant="ghost">{{ __('Cancel') }}</flux:button>
                         </flux:modal.close>
-                        <flux:button type="submit" variant="primary">{{ __('Create feature') }}</flux:button>
+                        <flux:button wire:click="deleteProject" variant="danger" icon="trash">{{ __('Delete project') }}</flux:button>
                     </div>
-                </form>
+                </div>
             </flux:modal>
-
-            @if ($this->canDeleteProject())
-                <flux:modal name="delete-project-modal" class="md:w-96">
-                    <div class="flex flex-col gap-4">
-                        <flux:heading size="lg">{{ __('Delete project?') }}</flux:heading>
-                        <flux:text>{{ __('This permanently removes the project, its features, stories, tasks, subtasks, approvals, and repo attachments. This cannot be undone.') }}</flux:text>
-                        <flux:input
-                            wire:model="deleteConfirmationName"
-                            :label="__('Type the project name to confirm')"
-                            :placeholder="$this->project->name"
-                            required
-                        />
-                        <div class="flex justify-end gap-2">
-                            <flux:modal.close>
-                                <flux:button type="button" variant="ghost">{{ __('Cancel') }}</flux:button>
-                            </flux:modal.close>
-                            <flux:button wire:click="deleteProject" variant="danger" icon="trash">{{ __('Delete project') }}</flux:button>
-                        </div>
-                    </div>
-                </flux:modal>
-            @endif
         @endif
     @endif
 </div>
